0.13.0.1

+++ CEX/ParallelUtils.cpp
@@ -3,7 +3,7 @@
 
 #if defined(_OPENMP)
 #	include <omp.h>
-#elif defined(_WIN32)
+#elif defined(CEX_OS_WINDOWS)
 #	include <Windows.h>
 #	include <ppl.h>
 #else
@@ -12,28 +12,28 @@
 
 NAMESPACE_UTILITY
 
-int ParallelUtils::ProcessorCount()
+size_t ParallelUtils::ProcessorCount()
 {
 #if defined(_OPENMP)
-	return omp_get_num_procs();
-#elif defined(_WIN32)
+	return static_cast<size_t>(omp_get_num_procs());
+#elif defined(CEX_OS_WINDOWS)
 	SYSTEM_INFO sysinfo;
 	GetSystemInfo(&sysinfo);
-	return sysinfo.dwNumberOfProcessors;
+	return static_cast<size_t>(sysinfo.dwNumberOfProcessors);
 #else
-	return std::thread::hardware_concurrency();
+	return static_cast<size_t>(std::thread::hardware_concurrency());
 #endif
 }
 
 void ParallelUtils::ParallelFor(size_t From, size_t To, const std::function<void(size_t)> &F)
 {
 #if defined(_OPENMP)
-#pragma omp parallel num_threads(To)
+#pragma omp parallel num_threads((int)To)
 	{
 		size_t i = (size_t)omp_get_thread_num();
 		F(i);
 	}
-#elif defined(_WIN32)
+#elif defined(CEX_OS_WINDOWS)
 	concurrency::parallel_for(From, To, [&](size_t i)
 	{
 		F(i);
